+++ PgRoutiner/SettingsManagement/Settings.cs
@@ -218,6 +218,8 @@ namespace PgRoutiner.SettingsManagement
         public bool MdIncludeSourceLinks { get; set; } = false;
         public bool MdIncludeExtensionLinks { get; set; } = false;
         public string MdSourceLinkRoot { get; set; } = null;
+        public bool MdIncludeTableCountEstimates { get; set; } = false;
+        public bool MdIncludeTableStats { get; set; } = false;
         public bool CommitMd { get; set; } = false;
 
         /*psql interactive terminal*/